Stripping HOST from Connect even in transparent mode.
Some websites (i.e. "nytimes.com") seem to have issues loading when using
LittleProxy in transparent mode.

The reason is down to the "URI" header containing the full URL (instead of just
the path) when using in Transparent mode.
While most webservers online are OK with that, some old servers seem to respond
with malformed "Location" headers when the Proxy is behaving in
"transparent mode" and it doesn't strip the Host from the URI.

So we are going to always strip the Host from the URI (transparent or not) and
ensure the Host header is present (added if missing).

This should ensure preservetion of correct behaviour with old clients and
old servers.