Shutdown Hook switch + UncaughtExceptionHandler.
So far, LittleProxy registered 2 important things
automatically, relying on the assumption that only
1, or just few, Proxy instances would be created
with a process. And that the end of the hosting
process would mean also the end of LittleProxy
instances.

This was translated with “hardcoded” shutdown hooks
registered on the Runtime, as well as UncaughtExceptionHander
for the Proxy Thread pool.

The issue with those are 1) making it impossible
for the host project to handle the lifecycle
fully 2) a memory leak building over time, when
multiple instances of LittleProxy got created and
destroyed in the lifetime of the host application.

Now it’s possible to control this aspect.